Check for Invalid HomeID or Controller Node ID's
ModemManager on the Host is known to interfere with OZW, and cause problems when starting up. Users will often get messages about "Out of Frame Flow" in OZW_Log.txt. See https://github.com/home-assistant/home-assistant/issues/21165 for example. 

One of the more bizarre side effects of this, is that it causes issues with OZW initializing and a side effect is Node 255 (Broadcast) address will be added to all Associations on devices. This causes weird issues such as a switch triggering every Z-Wave Node in the network etc. 

This check will bail out if we receive a invalid HomeID or Controller Node ID when starting up and stop OZW from inadvertently adding this broadcast address to all node associations.